pyproject.toml
[tool.poetry]
name = "vindex"
version = "0.1.0dev1"
description = "Vindex is a Discord bot made for DCS communities."
authors = ["Julien Mauroy <pro.julien.mauroy@gmail.com>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.3.2"
jishaku = "^2.5.2"
prisma = "^0.11.0"
click = "^8.1.7"
clickloader = "^0.2.0"
rich = "^13.7.0"
platformdirs = "^4.1.0"
pydantic = "^2.5.3"
uvloop = "^0.19.0"
babel = "^2.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
ruff = "^0.1.9"
poethepoet = "^0.24.4"
pylint = "^3.0.3"

[tool.poetry.scripts]
vindex = "vindex.cli:cli"

[tool.poe]
poetry_command = ""

[tool.poe.tasks]
_black = "black src/"
_isort = "isort src/"
_prisma_format = "prisma format"
format = ["_black", "_isort", "_prisma_format"]
lint = "pylint src --rcfile=.pylintrc --output-format=colorized"

[tool.black]
line-length = 99

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
